add in snapshot for dots and refactor use dot animation
