+++ packages/ui/src/components/breath-player/center-text.test.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import { create } from 'react-test-renderer'
+import { ColorThemeProvider } from '../../hooks/use-theme'
+import { CenterText } from './center-text'
+
+test('CenterText component matches snapshot', () => {
+  const tree = create(
+    <ColorThemeProvider value="dark">
+      <CenterText>Foo</CenterText>
+    </ColorThemeProvider>
+  ).toJSON()
+  expect(tree).toMatchSnapshot()
+})
